fix currency field display
On a Nothing Phone 1, on the Add transaction screen the currency were
not visible because the left and right paddings added too much empty
space to the text field and it just scrolled out from the view.

Removed the the padding from the currency field as the currency is
already aligend to the center. The padding is not needed, the spacing
will be evenly distributed before and after the text.