better display for 3char currency in posting row
The currency was too narrow to fit a 3 char currency there and the text
was scrolled out. To fix, increase the size of the currency field.
Also change how text weights defined:
Based on https://developer.android.com/jetpack/compose/modifiers#weight-in-row-and-column
it's not a percentage based calculation, but more like a ratio.