Tweak posting row item widths to make currency a bit wider (#62)
* fix currency field display

On a Nothing Phone 1, on the Add transaction screen the currency were
not visible because the left and right paddings added too much empty
space to the text field and it just scrolled out from the view.

Removed the the padding from the currency field as the currency is
already aligend to the center. The padding is not needed, the spacing
will be evenly distributed before and after the text.

* better display for 3char currency in posting row

The currency was too narrow to fit a 3 char currency there and the text
was scrolled out. To fix, increase the size of the currency field.
Also change how text weights defined:
Based on https://developer.android.com/jetpack/compose/modifiers#weight-in-row-and-column
it's not a percentage based calculation, but more like a ratio.

* Slightly tweak weights

---------

Co-authored-by: Charlotte Van Petegem <charlotte@vanpetegem.me>